+++ howfairis/__init__.py
@@ -1,5 +1,6 @@
 from .__version__ import __version__
 from .checker import Checker
+from .cli import cli
 from .code_repository_platforms import Platform
 from .compliance import Compliance
 from .repo import Repo
@@ -11,6 +12,7 @@ __email__ = 'j.spaaks@esciencecenter.nl'
 __all__ = [
     "__version__",
     "Checker",
+    "cli",
     "Compliance",
     "Platform",
     "Repo"
