+++ howfairis/cli/__init__.py
@@ -1,0 +1,6 @@
+from .cli import cli
+
+
+__all__ = [
+    "cli"
+]
