+++ howfairis/cli/exit_with_default_config.py
@@ -1,0 +1,13 @@
+import os
+import sys
+
+
+def exit_with_default_config(is_quiet=False):
+    if not is_quiet:
+        parent_dir = os.path.dirname(__file__)
+        pkg_root = os.path.join(parent_dir, "..")
+        default_config_filename = os.path.join(pkg_root, "data", ".howfairis.yml")
+        with open(default_config_filename, "rt") as f:
+            text = f.read()
+        print(text)
+    sys.exit(0)
