+++ howfairis/cli/print_feedback_about_config_args.py
@@ -1,0 +1,14 @@
+def print_feedback_about_config_args(ignore_repo_config, repo_config_filename, user_config_filename, is_quiet=False):
+
+    if ignore_repo_config is True and is_quiet is False:
+        print("Ignoring any configuration files on the remote.")
+
+    if ignore_repo_config is True:
+        assert repo_config_filename is None, "When ignoring any configuration files on the remote, you" + \
+                                             " should not set a remote configuration filename."
+
+    if ignore_repo_config is False and repo_config_filename is not None and is_quiet is False:
+        print("Remote configuration filename: " + repo_config_filename)
+
+    if user_config_filename is not None:
+        print("Local configuration file: " + user_config_filename)
