+++ setup.cfg
@@ -31,3 +31,5 @@ lines_after_imports = 2
 force_single_line = 1
 no_lines_before = FUTURE,STDLIB,THIRDPARTY,FIRSTPARTY,LOCALFOLDER
 known_first_party = howfairis
+src_paths = howfairis,tests,livetests
+line_length = 120
